requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.9
openai>=1.3.0
chromadb>=0.4.18
tiktoken>=0.5.1
faiss-cpu>=1.7.4
unstructured>=0.10.30
pypdf>=3.17.1
sqlalchemy>=2.0.23
pandas>=2.1.3
matplotlib>=3.8.2
plotly>=5.18.0
streamlit>=1.28.2
yfinance>=0.2.31
fredapi>=0.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
lxml>=4.9.3 